iOS 添加native2flutter方法调用

+++ example/flutter_app/ios/Classes/MQQFlutterGen_/flutter2native/MQQFlutterGen_InnerClass.h
@@ -5,11 +5,17 @@ NS_ASSUME_NONNULL_BEGIN
 @interface MQQFlutterGen_InnerClass : NSObject <NSCopying>
 
 
+/// Note: this method create by SPI, if change Class property or method,
+/// please remove it. it will be carted by SPI again.
+/// Note: this method create by SPI, if change Class property or method,
+/// please remove it. it will be carted by SPI again.
+/// Note: this method create by SPI, if change Class property or method,
+/// please remove it. it will be carted by SPI again.
 /// Dart property declaration: String? a;.
 @property (nonatomic, strong, nullable) NSString *a;
 
 /// Dart property declaration: int? b;.
-@property (nonatomic, assign) int b;
+@property (nonatomic, assign) long long b;
 
 
 @end
